Integration test automation
Start of adding integration tests to the XCTest target (i.e. tests that run against a live server, not mock).

This commit adds the main infrastructure for integration test automation: signing in as admin user for creating/deleting test accounts (while keeping admin credentials *out* of source control), and being able to sign in to those test accounts without signing out from the auth account.

Also created full integration tests for auth manager signUp/signIn/signOut methods. In the process found a bug in the signOut method and fixed it.

Todo: Tests for the rest of it.