Clear user info from cache before sign-in
because UserProfile and DataGroups have client-writable fields, they won’t update from the server once cached; and the reason for having to sign in might be because the session was invalidated when the researcher UI was used to edit the StudyParticipant (from which the deprecated UserProfile and DataGroups objects are now generated for backward compatibility).

When StudyParticipant support is implemented in this SDK, we’ll need to clear that from cache before sign-in as well.