Fix reauth upload spam
Dwayne was occasionally seeing bursts of upload-request-fails-with-401/app-reauth/retry-fails-with-401/rinse/repeat. I finally saw this repro’ed when testing a fix for Sarcoidosis. The culprit was that for background requests, 401s were being retried after reauth without updating the session token in the headers of the request being retried. Derp.