tweaks and tests
CAVEAT: testUpdateScheduledActivities integration test failing as of today (2017-04-19) because of a bug in the server implementation where this endpoint returns the scheduledOn and expiresOn times for ScheduledActivity objects as the date and time-of-day but with the time zone specifier set to Z instead of the local time zone of the request. Alx will fix on the server.