redo the placeholder UserSessionInfo
was running into the same issue as the recently-fixed unit test crasher, namely, without credentials we can’t decrypt StudyParticipant, so when we try to get its NSObject doppelgänger (SBBStudyParticipant) from cache we get nil even if it exists. So instead we’ll just create one outside of cache and give it to the auth delegate to use until we get a real one into cache.