Fix crash caused by race condition in context init
Because the method for initializing the `cacheIOContext` first checks to see if the context is nil and returns that context if not, need to set the ivar *after* the persistent store is set up. Otherwise, there is a race condition where the context can be returned before the persistent store model has been attached.