Use /v4 endpoint to save localDate reports
Turns out you can’t actually mix-and-match saving reports with the v4 endpoint and then retrieving them with the v3 one. Which is fine since it also turns out the v3 endpoint doesn’t offer the convenience I thought it did with not requiring paging—it’s just that the 45-day retrieval limit wasn’t in the Swagger docs.