bump integration test timeout
Alx insists the server response times (on the order of 1 sec at worst) are nowhere near as bad as the long network response delays I’m seeing in these tests sometimes (>30 sec), and I have gigabit fiber at home, so it must be something in the Simulator’s networking or the internets in general that is intermittently super slow to pass things on to the app.

Also, disabling one test temporarily due to changes in the non-client-facing API used to set it up on the server.